Split customers example into smaller files
