+++ src/App/Documentation/examples/customers/index.js
@@ -1,14 +1,13 @@
-import React, { Component, useEffect } from "react";
+import React, { Component } from "react";
 import PropTypes from "prop-types";
-import { DocContainer, ComponentPreview } from "@docutils";
+import { DocContainer } from "@docutils";
 
-import PaginationComponent from "@components/Pagination";
-import MediaObjectComponent from "@components/MediaObject";
-import { Datepicker as DatepickerComponent, Togglebox, Checkbox } from "@components/FormComponents";
-import StepsComponent from "@components/Steps";
-import ActionLinkComponent from "@components/ActionLink";
-import InputGroup from "@components/InputGroup";
-import Chart from "@components/Chart";
+import CustomersOverview from "./CustomersOverview";
+import CustomersDetailed from "./CustomersDetailed";
+import CustomersDetailedOrders from "./CustomersDetailed/components/CustomersDetailedOrders";
+import CustomersDetailedInquiries from "./CustomersDetailed/components/CustomersDetailedInquiries";
+import CustomersDetailedCharts from "./CustomersDetailed/components/CustomersDetailedCharts";
+import CustomersDetailedSettings from "./CustomersDetailed/components/CustomersDetailedSettings";
 
 import {
     customersList,
@@ -20,549 +19,13 @@ import {
     customersDetailedPreviousInquiries
 } from "./constants";
 
-const { actionList, datepicker, dialog, tabs } = window.dg;
+const { actionList } = window.dg;
 
 const createNumArray = (length, baseNum, addNum) => (
     [...Array(length)].map(() => (
         Math.floor(Math.random() * addNum) + Math.floor(Math.random() * addNum) + Math.floor(Math.random() * addNum) +
         Math.floor(Math.random() * addNum) + Math.floor(Math.random() * addNum) + baseNum)
     ) // The multiple Math.floor(Math.random) is there to achieve a lower variance
-);
-
-class CustomersOverview extends Component {
-    constructor (props) {
-        super(props);
-        this.state = {
-            customerFilter: ""
-        };
-    }
-
-    componentDidUpdate () {
-        this.customerFilterField.focus();
-    }
-
-    updateFilter (e) {
-        this.setState({
-            customerFilter: e.target.value.toLowerCase()
-        });
-    }
-
-    resetFilter () {
-        this.setState({
-            customerFilter: ""
-        });
-    }
-
-    render () {
-        return (
-            <>
-                <h2 id="customers-overview">Customers overview</h2>
-                <ComponentPreview language="html" showCasePanel codeFigure>
-                    <div className="panel panel-brand">
-                        <header>
-                            <div className="row">
-                                <div className="col-lg-4">
-                                    <h2 className="panel-title">Customers</h2>
-                                </div>
-                                <div className="col-lg-8">
-                                    <div className="input-group">{"\n"}
-                                        <span className="input-group-addon">{"\n"}
-                                            <i className="material-icons">search</i>
-                                        </span>{"\n"}
-                                        <input ref={input => this.customerFilterField = input} value={this.state.customerFilter} onChange={e => this.updateFilter(e)} type="text" className="form-control" />{"\n"}
-                                        <button type="button" onClick={() => this.resetFilter()} className="btn btn-executive">{"\n"}
-                                            <i className="material-icons">close</i>{"\n"}
-                                        </button>{"\n"}
-                                    </div>
-                                </div>
-                            </div>
-                        </header>
-                        <div className="panel-table">
-                            <table className="table table-hover table-striped">
-                                <thead>
-                                    <tr>
-                                        <th scope="col">ID</th>
-                                        <th scope="col" className="d-none d-sm-table-cell">First name</th>
-                                        <th scope="col" className="d-none d-sm-table-cell">Last name</th>
-                                        <th scope="col">Location</th>
-                                        <th scope="col" className="d-none d-lg-table-cell">Delivery status</th>
-                                        <th scope="col" className="d-lg-none" />
-                                    </tr>
-                                </thead>
-                                <tbody>
-                                    {customersList.filter(
-                                        customer => this.state.customerFilter === "" ||
-                                        customer.firstName.toLowerCase().includes(this.state.customerFilter) ||
-                                        customer.lastName.toLowerCase().includes(this.state.customerFilter) ||
-                                        customer.location.toLowerCase().includes(this.state.customerFilter)
-                                    )
-                                        .map((customer, index) => (
-                                            <tr key={customer.id} onClick={() => this.props.setCustomerIndex(index)}>
-                                                <th scope="row">{customer.id}</th>
-                                                <td className="d-none d-sm-table-cell">{customer.firstName}</td>
-                                                <td className="d-none d-sm-table-cell">{customer.lastName}</td>
-                                                <td>{customer.location}</td>
-                                                <td className="d-none d-lg-table-cell">
-                                                    <span className={`status status-${customer.status}`}>{statusText[customer.status]}</span>
-                                                </td>
-                                                <td className="d-lg-none">
-                                                    <span className={`status status-${customer.status} pl-0`} />
-                                                </td>
-                                            </tr>
-                                        ))}
-                                </tbody>
-                            </table>
-                        </div>
-                        <PaginationComponent text="1 of 123" arrows farArrows />
-                    </div>
-                </ComponentPreview>
-            </>
-        );
-    }
-}
-
-class CustomersDetailed extends Component {
-    constructor (props) {
-        super(props);
-        this.state = {
-            tabIndex: 0,
-            tabScrollState: null
-        };
-    }
-
-    componentDidUpdate () {
-        actionList.init();
-        dialog.init();
-        tabs.init();
-        tabs.setScrollState("customers-detailed-tabs", this.state.tabScrollState);
-    }
-
-    componentDidMount () {
-        dialog.init();
-        tabs.init();
-    }
-
-    selectTab (e, i) {
-        e.preventDefault();
-
-        const tabScroll = document.getElementById("customers-detailed-tabs-ul");
-        const scrollStart = tabScroll.scrollLeft;
-        const scrollTotalAmount = (e.target.offsetLeft - (tabScroll.offsetWidth / 2) + (e.target.offsetWidth / 2)) - scrollStart;
-
-        this.setState({
-            tabIndex: i,
-            tabScrollState: {
-                scrollTotalAmount,
-                scrollStart
-            }
-        });
-    }
-
-    render () {
-        return (
-            <>
-                <h2 id="customers-detailed">Customers detailed</h2>
-                <ComponentPreview language="html" showCasePanel codeFigure>
-                    <div className="panel panel-brand">
-                        <header>
-                            <ol className="breadcrumb">
-                                <li><a href="#">Customers</a></li>
-                                <li><span>{this.props.customer.id}</span></li>
-                            </ol>
-                            <div className="row">
-                                <div className="col-auto mr-auto">
-                                    <div className="row">
-                                        <div className="col-xs-auto">
-                                            <div className="d-none d-sm-block">
-                                                <div className="media media-sm">
-                                                    <div className="media-img">
-                                                        <i className="material-icons">account_circle</i>
-                                                    </div>
-                                                    <div className="media-body">
-                                                        <h4>{`${this.props.customer.firstName} ${this.props.customer.lastName}`}</h4>
-                                                        <p>{`${this.props.customer.email} ${this.props.customer.phone}`}</p>
-                                                    </div>
-                                                </div>
-                                            </div>
-                                            <div className="d-block d-sm-none">
-                                                <div className="media media-sm">
-                                                    <div className="media-img">
-                                                        <i className="material-icons">account_circle</i>
-                                                    </div>
-                                                    <div className="media-body">
-                                                        <h4>{`${this.props.customer.firstName} ${this.props.customer.lastName}`}</h4>
-                                                    </div>
-                                                </div>
-                                            </div>
-                                        </div>
-                                        <div className="col-xs-auto">
-                                            <div id="customer-detailed-action-list" className="action-list anchor-top-right">
-                                                <button id="actionListToggle" type="button" className="action-toggle" aria-haspopup="true">
-                                                    <i className="material-icons">more_vert</i>
-                                                </button>
-                                                <div className="action-menu" aria-labelledby="actionListToggle">
-                                                    <a href="#" onClick={e => e.preventDefault()}><i className="material-icons">edit</i>Edit</a>
-                                                    <a href="#" onClick={e => e.preventDefault()} data-dialog-open="customers-detailed-dialog"><i className="material-icons">delete</i>Delete</a>
-                                                </div>
-                                            </div>
-                                        </div>
-                                        <CustomersDetailedDialog customer={this.props.customer} />
-                                    </div>
-                                    <div className="row d-block d-sm-none">
-                                        <div className="col-sm-auto pl-0">
-                                            <p className="mb-0 text-truncate">{this.props.customer.email}</p>
-                                        </div>
-                                        <div className="col-sm-auto pl-0">
-                                            <p>{this.props.customer.phone}</p>
-                                        </div>
-                                    </div>
-                                </div>
-                                <div className="col-auto pl-0">
-                                    <h4>Delivery status:</h4>
-                                    <span className={`status status-${this.props.customer.status}`}>{statusText[this.props.customer.status]}</span>
-                                </div>
-                            </div>
-                        </header>
-                        <div className="panel-body">
-                            <div id="customers-detailed-tabs" className="tabs tabs-scroll">{"\n"}
-                                <ul id="customers-detailed-tabs-ul">
-                                    {this.props.customersDetailedTabs.map((tab, i) => (
-                                        <li key={`tab-item-${i}`} className={this.state.tabIndex === i ? "active" : null}>{"\n"}
-                                            <a href="#" onClick={e => this.selectTab(e, i)}>{tab.name}</a>{"\n"}
-                                        </li>
-                                    ))}
-                                </ul>
-                            </div>
-
-                            {this.props.customersDetailedTabs[this.state.tabIndex].component}
-                        </div>
-                    </div>
-                </ComponentPreview>
-            </>
-        );
-    }
-
-}
-
-const CustomersDetailedDialog = ({ customer }) => (
-    <div className="dialog" id="customers-detailed-dialog" role="dialog" aria-modal="true" aria-labelledby="aria-label-dia" aria-describedby="aria-describe-dia">
-        <section>
-            <header className="dialog-header" id="aria-label-dia">
-                <h4>Delete customer {customer.id} {customer.firstName} {customer.lastName}?</h4>
-                <button type="button" className="dialog-close">
-                    <i className="material-icons">close</i>
-                </button>
-            </header>
-            <div className="dialog-body" id="aria-describe-dia">
-                <p>Are you sure you want to permanently delete the customer {customer.id} {customer.firstName} {customer.lastName}?</p>
-            </div>
-            <footer className="dialog-footer">
-                <button className="btn btn-guiding" type="button" data-dialog-close="true">Cancel</button>
-                <button className="btn btn-destructive" type="button">Delete</button>
-            </footer>
-        </section>
-    </div>
-);
-
-const CustomersDetailedDatePickerGroup = () => {
-
-    useEffect(() => {
-        datepicker.init();
-    });
-
-    return (
-        <div className="row">
-            <div className="col-sm-6">
-                <DatepickerComponent
-                    value="2000-01-01"
-                    label="From:"
-                    prefixType="icon"
-                    prefixValue="event"
-                    fulldate
-                    id="init-value-datepicker"
-                />
-            </div>
-            <div className="col-sm-6">
-                <DatepickerComponent
-                    value="2020-01-01"
-                    label="To:"
-                    prefixType="icon"
-                    prefixValue="event"
-                    fulldate
-                    id="init-value-datepicker"
-                />
-            </div>
-        </div>
-    );
-};
-
-const CustomersDetailedOrders = ({ customersDetailedOrdersSteps, customersDetailedOrders }) => (
-    <>
-        <h3>Latest order</h3>
-        <div className="row">
-            <div className="col-md-6">
-                <div className="slab border slab-elevated">
-                    <h5>Order: OrderID#123456</h5>
-                    <p>Price: £ 1994</p>
-                    <p>Date: 22.10.2019</p>
-                    <p>Product ID: ProductID#321</p>
-                    <p>Product name: Product name ABC</p>
-                    <p>Addtional information: Description of the product and/or additional comments from the customer with regards to the order or product.</p>
-                    <ActionLinkComponent linkText="Edit order" smallText="OrderID#123456" multiline={true} />
-                </div>
-            </div>
-            <div className="col-md-6 mt-3 mt-md-0">
-                <h5>Detailed internal status:</h5>
-                <StepsComponent steps={customersDetailedOrdersSteps} vertical />
-            </div>
-        </div>
-
-        <h3>Previous orders</h3>
-        <CustomersDetailedDatePickerGroup />
-        {customersDetailedOrders.map((order, i) => (
-            <div key={i} className="slab border slab-elevated">
-                <div className="d-flex">
-                    <div className="flex-grow-1">
-                        <div className="flex-column">
-                            <p className="font-weight-bold">
-                                {`OrderID#${i}: `}
-                            </p>
-                            <p>
-                                {`ProductID#${i * 5} Product name XYZ${i}`}
-                            </p>
-                        </div>
-                    </div>
-                    <div>
-                        <p className="font-weight-bold text-nowrap">£ {order}</p>
-                    </div>
-                </div>
-                <ActionLinkComponent linkText="Edit order" smallText={`OrderID#${i}:`} multiline={true} />
-            </div>
-        ))}
-    </>
-);
-
-const CustomersDetailedInquiryCard = ({ inquiry, size }) => (
-    <div className="card card-secondary">
-        <header className="card-header">
-            <h4>{`Inquiry: ${inquiry.id} | ${inquiry.date}`}</h4>
-        </header>
-        <div className="card-body">
-            <div className="d-none d-sm-block">
-                <div className="d-flex justify-content-around">
-                    <MediaObjectComponent
-                        size={size}
-                        icon="message"
-                        heading={inquiry.numMessages.toString()}
-                        text="messages"
-                    />
-
-                    <MediaObjectComponent
-                        size={size}
-                        icon={inquiry.resolved ? "check_circle" : "feedback"}
-                        heading={inquiry.resolved ? "Resolved" : "Unresolved"}
-
-                    />
-                </div>
-            </div>
-            <div className="d-sm-none">
-                <div className="d-flex flex-column">
-                    <MediaObjectComponent
-                        size="sm"
-                        icon="message"
-                        heading={inquiry.numMessages.toString()}
-                        text="messages"
-                    />
-
-                    <MediaObjectComponent
-                        size="sm"
-                        icon={inquiry.resolved ? "check_circle" : "feedback"}
-                        heading={inquiry.resolved ? "Resolved" : "Unresolved"}
-
-                    />
-                </div>
-            </div>
-            <h5>Initial message:</h5>
-            <p>{inquiry.message}</p>
-        </div>
-        <div className="card-footer">
-            <div className="footer-link"><a href="#">See inquiry thread</a></div>
-        </div>
-    </div>
-);
-
-const CustomersDetailedInquiries = ({ customersDetailedLatestInquiry, customersDetailedPreviousInquiries }) => (
-    <>
-        <h3>Latest inquiry</h3>
-        <CustomersDetailedInquiryCard inquiry={customersDetailedLatestInquiry} size="lg" />
-
-        <h3>Previous inquiries</h3>
-        <CustomersDetailedDatePickerGroup />
-        <div className="row">
-            {customersDetailedPreviousInquiries.map(inquiry => (
-                <div key={inquiry.id} className="col-lg-6 d-flex">
-                    <CustomersDetailedInquiryCard inquiry={inquiry} size="sm" />
-                </div>
-            ))}
-        </div>
-    </>
-);
-
-const CustomersDetailedCharts = ({ customerIdName, customerOrders, customerInquiries, createNumArray }) => {
-    const datasetDataLength = customerOrders.length;
-    const totalNumMessages = customerInquiries.reduce((accumulator, inquiry) => accumulator + inquiry.numMessages, 0);
-
-    return (
-        <>
-            <h3>Orders</h3>
-            <h4>Historical purchase prices</h4>
-            <Chart
-                id="customers-detailed-charts-1"
-                options={{
-                    type: "bar",
-                    data: {
-                        labels: [...Array(datasetDataLength).keys()].map(i => (2019 - (datasetDataLength - 1) + i).toString()),
-                        datasets: [
-                            {
-                                label: customerIdName,
-                                data: customerOrders,
-                                fill: false
-                            },
-                            {
-                                label: "Customers average",
-                                data: createNumArray(datasetDataLength, 1000, 200),
-                                fill: false
-                            },
-                            {
-                                label: "Estimated",
-                                data: createNumArray(datasetDataLength, 850, 200),
-                                fill: false
-                            }
-                        ]
-                    },
-                    options: {
-                        scales: {
-                            yAxes: [{
-                                scaleLabel: {
-                                    display: true,
-                                    labelString: "British Pound (£)"
-                                }
-                            }]
-                        }
-                    }
-                }}
-            />
-
-            <h3>Inquiries</h3>
-            <h4>Number of inquiry messages</h4>
-            <Chart
-                id="customers-detailed-charts-2"
-                options={{
-                    type: "line",
-                    data: {
-                        labels: customerInquiries.map(inquiry => inquiry.date),
-                        datasets: [
-                            {
-                                label: customerIdName,
-                                data: customerInquiries.map(inquiry => inquiry.numMessages),
-                                fill: false
-                            },
-                            {
-                                label: "Customers average",
-                                data: createNumArray(customerInquiries.length, 3, 4), // The multiple Math.floor(Math.random) is there to achieve a less random looking distribution
-                                fill: false
-                            }
-                        ]
-                    },
-                    options: {
-                        scales: {
-                            yAxes: [{
-                                ticks: {
-                                    beginAtZero: true
-                                }
-                            }]
-                        }
-                    }
-                }}
-            />
-            <div className="row">
-                <div className="col-lg-6">
-                    <h4>Status distribution</h4>
-                    <Chart id="customers-detailed-charts-3"
-                        options={{
-                            type: "pie",
-                            data: {
-                                labels: ["Resolved", "Unresolved", "Unknown"],
-                                datasets: [
-                                    {
-                                        data: [customerInquiries.filter(inquiry => inquiry.resolved).length, customerInquiries.filter(inquiry => !inquiry.resolved).length, 1]
-                                    }
-                                ]
-                            }
-                        }}
-                    />
-                </div>
-                <div className="col-lg-6">
-                    <h4>Time before response</h4>
-                    <Chart id="customers-detailed-charts-4"
-                        options={{
-                            type: "pie",
-                            data: {
-                                labels: ["< 2 days", "< 1 week", "< 1 month", "< 3 months", "Other"],
-                                datasets: [
-                                    {
-                                        data: [
-                                            Math.floor(totalNumMessages * 0.4),
-                                            Math.floor(totalNumMessages * 0.3),
-                                            Math.floor(totalNumMessages * 0.15),
-                                            Math.floor(totalNumMessages * 0.1),
-                                            Math.floor(totalNumMessages * 0.05)
-                                        ]
-                                    }
-                                ]
-                            }
-                        }}
-                    />
-                </div>
-            </div>
-
-        </>
-    );
-};
-
-const CustomersDetailedSettings = () => (
-    <>
-        <h3>Notifications</h3>
-        <div className="row justify-content-around mt-3">
-            <div className="col-lg-auto m-2">
-                <div className="d-flex flex-column">
-                    <Togglebox id="customers-detailed-settings-toggle-1" checked label="Delivery status changes" />
-                    <Togglebox id="customers-detailed-settings-toggle-3" label="Critical delivery messages" />
-                </div>
-            </div>
-            <div className="col-lg-auto m-2">
-                <Togglebox id="customers-detailed-settings-toggle-2" label="New inquiry messages" />
-                <Togglebox id="customers-detailed-settings-toggle-4" checked label="Inquiry status changes" />
-            </div>
-        </div>
-        <h3>Newsletter and information</h3>
-        <div className="row mt-4">
-            {[...Array(7).keys()].map(item => (
-                <div key={item} className="col-lg-3 col-md-4 col-sm-6">
-                    <Checkbox id={`customers-detailed-settings-checkbox-${item}`} checked label={`Subscription #${item + 1}`} />
-                </div>
-            ))}
-        </div>
-
-        <h3>Customer type</h3>
-        <InputGroup
-            type="select"
-            selectOptions={["Long-term contract", "Short-term contract", "One time purchase"]}
-            prefixValue="Type"
-        />
-
-        <button className="btn btn-executive">Save</button>
-    </>
 );
 
 class Customers extends Component {
@@ -642,11 +105,16 @@ class Customers extends Component {
 
         return (
             <DocContainer>
-                <CustomersOverview setCustomerIndex={customerIndex => this.setCustomerIndex(customerIndex)}/>
+                <CustomersOverview
+                    setCustomerIndex={customerIndex => this.setCustomerIndex(customerIndex)}
+                    customersList={customersList}
+                    statusText={statusText}
+                />
                 <CustomersDetailed
                     customer={customersList[this.state.customerIndex]}
                     customersDetailedTabs={customersDetailedTabs}
                     customersDetailedActionList={customersDetailedActionList}
+                    statusText={statusText}
                 />
             </DocContainer>
         );
@@ -657,87 +125,4 @@ Customers.propTypes = {
     test: PropTypes.bool
 };
 
-CustomersOverview.propTypes = {
-    setCustomerIndex: PropTypes.func.isRequired
-};
-
-CustomersDetailed.propTypes = {
-    customer: PropTypes.exact({
-        id: PropTypes.string.isRequired,
-        firstName: PropTypes.string.isRequired,
-        lastName: PropTypes.string.isRequired,
-        email: PropTypes.string.isRequired,
-        phone: PropTypes.string.isRequired,
-        location: PropTypes.string.isRequired,
-        status: PropTypes.string.isRequired
-    }).isRequired,
-    customersDetailedTabs: PropTypes.arrayOf(PropTypes.exact({
-        name: PropTypes.string.isRequired,
-        component: PropTypes.object.isRequired
-    })).isRequired
-};
-
-CustomersDetailedOrders.propTypes = {
-    customersDetailedOrdersSteps: PropTypes.arrayOf(PropTypes.shape({
-        title: PropTypes.string.isRequired,
-        ongoing: PropTypes.bool,
-        completed: PropTypes.bool,
-        selected: PropTypes.bool
-    })).isRequired,
-    customersDetailedOrders: PropTypes.arrayOf(PropTypes.number).isRequired
-};
-
-CustomersDetailedInquiries.propTypes = {
-    customersDetailedLatestInquiry: PropTypes.exact({
-        id: PropTypes.string.isRequired,
-        date: PropTypes.string.isRequired,
-        numMessages: PropTypes.number.isRequired,
-        resolved: PropTypes.bool.isRequired,
-        message: PropTypes.string.isRequired
-    }).isRequired,
-    customersDetailedPreviousInquiries: PropTypes.arrayOf(PropTypes.exact({
-        id: PropTypes.string.isRequired,
-        date: PropTypes.string.isRequired,
-        numMessages: PropTypes.number.isRequired,
-        resolved: PropTypes.bool.isRequired,
-        message: PropTypes.string.isRequired
-    })).isRequired
-};
-
-CustomersDetailedCharts.propTypes = {
-    customerIdName: PropTypes.string.isRequired,
-    customerOrders: PropTypes.arrayOf(PropTypes.number).isRequired,
-    customerInquiries: PropTypes.arrayOf(PropTypes.exact({
-        id: PropTypes.string.isRequired,
-        date: PropTypes.string.isRequired,
-        numMessages: PropTypes.number.isRequired,
-        resolved: PropTypes.bool.isRequired,
-        message: PropTypes.string.isRequired
-    })).isRequired
-};
-
-CustomersDetailedDialog.propTypes = {
-    customer: PropTypes.exact({
-        id: PropTypes.string.isRequired,
-        firstName: PropTypes.string.isRequired,
-        lastName: PropTypes.string.isRequired,
-        email: PropTypes.string.isRequired,
-        phone: PropTypes.string.isRequired,
-        location: PropTypes.string.isRequired,
-        status: PropTypes.string.isRequired
-    }).isRequired
-};
-
 export default Customers;
-
-export {
-    CustomersOverview,
-    CustomersDetailed,
-    CustomersDetailedDatePickerGroup,
-    CustomersDetailedOrders,
-    CustomersDetailedInquiries,
-    CustomersDetailedInquiryCard,
-    CustomersDetailedSettings,
-    CustomersDetailedCharts,
-    CustomersDetailedDialog
-};
